move download function to utils
